+++ src/test/java/de/alpharogroup/collections/array/ArrayFactoryTest.java
@@ -41,6 +41,20 @@ public class ArrayFactoryTest
 {
 
 	/**
+	 * Test for method {@link ArrayFactory#newArray(Class, int)}
+	 */
+	@Test
+	public void testNewArrayTypeWithLength()
+	{
+		int actual;
+		int expected;
+		final Integer[] array = ArrayFactory.newArray(Integer.class, 3);
+		actual = array.length;
+		expected = 3;
+		assertEquals(expected, actual);
+	}
+
+	/**
 	 * Test for method {@link ArrayFactory#newArray(Object...)}
 	 */
 	@Test
@@ -61,6 +75,21 @@ public class ArrayFactoryTest
 		expected = 0;
 		actual = objects.length;
 		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test for method {@link ArrayFactory#newEmptyArray(Object[])}
+	 */
+	@Test
+	public void testNewEmptyArray()
+	{
+		Integer[] actual;
+		Integer[] expected;
+		final Integer[] numbers = { 1, 2, 3 };
+		expected = new Integer[3];
+		actual = ArrayFactory.newEmptyArray(numbers);
+		assertTrue(Arrays.equals(actual, expected));
+
 	}
 
 	/**