+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 Version 5.3-SNAPSHOT
 -------------
 
+ADDED:
+
+- new factory method for create an array from a given class type and a given length
+- new factory method for create an empty array from array with the length of the given array
+- new factory method for create an empty array from array with a given length
+- new extension method created that removes a given prefix array from a given array
+- new extension method created that removes a given suffix array from a given array
+
 Version 5.2.1
 -------------